+++ src/store/module-ds/admin/actions-role.js
@@ -5,9 +5,17 @@ const getUrl = getters => getters.rolesUrl;
 const getPermissionsUrl = (getters, roleId) => getters.permissionsUrl(roleId);
 const setDsMutation = 'setDsRoles';
 const getAction = 'getRoles';
+const sysObject = 'role';          // название сист. объекта
+const sysObject2 = 'role.permission';         // название сист. объекта
 
 // получить полный источник данных
 export const getRoles = async ({ commit, getters }) => {
+  // проверка разрешений
+  if (!getters.isGranted('read', sysObject)) {
+    commit(setDsMutation, []);
+    return null;
+  }
+
   const url = getUrl(getters);
   const response = await axios.get(url)
     .then((resp) => { commit(setDsMutation, resp.data); commit('preparePermissionViewData'); return resp; })
@@ -20,9 +28,14 @@ export const getRoles = async ({ commit, getters }) => {
 };
 
 // добавить документ
-export const addRole = async ({ getters, dispatch }, obj) => {
-  const url = getUrl(getters);
+export const addRole = async ({ commit, getters, dispatch }, obj) => {
+  // проверка разрешений
+  if (!getters.isGranted('create', sysObject)) {
+    commit(setDsMutation, []);
+    return null;
+  }
 
+  const url = getUrl(getters);
   const postData = {
     name: obj.name,
     tag: obj.tag,
@@ -35,9 +48,14 @@ export const addRole = async ({ getters, dispatch }, obj) => {
 };
 
 // удалить документ
-export const deleteRole = async ({ getters, dispatch }, id) => {
-  const url = `${getUrl(getters)}/${id}`;
+export const deleteRole = async ({ commit, getters, dispatch }, id) => {
+  // проверка разрешений
+  if (!getters.isGranted('delete', sysObject)) {
+    commit(setDsMutation, []);
+    return null;
+  }
 
+  const url = `${getUrl(getters)}/${id}`;
   const response = await axios.delete(url);
   await dispatch(getAction);
   return response;
@@ -45,6 +63,12 @@ export const deleteRole = async ({ getters, dispatch }, id) => {
 
 // добавить встроенные документы разрешений
 export const addPermissions = async ({ getters, dispatch }, obj) => {
+  // проверка разрешений
+  if (!getters.isGranted('create', sysObject2)) {
+    // commit(setDsMutation, []);
+    return null;
+  }
+
   const url = getPermissionsUrl(getters, obj.roleId);
   const payload = {
     system_objectIds: obj.system_objectIds,
@@ -59,8 +83,13 @@ export const addPermissions = async ({ getters, dispatch }, obj) => {
 
 // изменить разрешение
 export const updatePermission = async ({ getters, dispatch }, obj) => {
-  const url = `${getPermissionsUrl(getters, obj.roleId)}/${obj.system_objectId}`;
+  // проверка разрешений
+  if (!getters.isGranted('update', sysObject2)) {
+    // commit(setDsMutation, []);
+    return null;
+  }
 
+  const url = `${getPermissionsUrl(getters, obj.roleId)}/${obj.system_objectId}`;
   const putData = {
     actionId: obj.actionId,
     granted: obj.granted,
@@ -74,8 +103,13 @@ export const updatePermission = async ({ getters, dispatch }, obj) => {
 
 // удалить разрешение
 export const deletePermission = async ({ getters, dispatch }, obj) => {
-  const url = `${getPermissionsUrl(getters, obj.roleId)}/${obj.system_objectId}`;
+  // проверка разрешений
+  if (!getters.isGranted('delete', sysObject2)) {
+    // commit(setDsMutation, []);
+    return null;
+  }
 
+  const url = `${getPermissionsUrl(getters, obj.roleId)}/${obj.system_objectId}`;
   const response = await axios.delete(url);
   await dispatch('getRoles');
   return response;
